+++ R/dbplyr.R
@@ -3,7 +3,6 @@
 #' @importFrom stats setNames update
 #' @importFrom utils head tail
 #' @importFrom glue glue
-#' @importFrom methods setOldClass
 #' @importFrom dplyr n
 #' @importFrom blob is_blob as_blob
 #' @import rlang