+++ NEWS.md
@@ -1,3 +1,5 @@
+# dbplyr (development version)
+
 # dbplyr 1.4.4
 
 * Internally `DBI::dbExecute()` now uses `immediate = TRUE`; this improves