+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * All old lazy eval shims have been removed. 
 
+# dbplyr 1.4.4
+
 * Internally `DBI::dbExecute()` now uses `immediate = TRUE`; this improves
   support for session-scoped temporary tables in MS SQL (@krlmlr, #438).